requirements.txt
fastapi
uvicorn[standard]
timm
torch
Pillow
loguru
